+++ Enums.cs
@@ -3,5 +3,6 @@
     public static class Enums
     {
         public enum eRopeVineChainOptions { Vanilla, AllowClimbing, AutoDismount }
+        public enum eHotSunCookingOptions { None, OnlyFood, AllCookables }
     }
 }